fix init in FIL triton and separate tests into separate folder for ops
